Integration testing

+++ respondent/models.py
@@ -3,7 +3,7 @@ import uuid
 from django.contrib.auth.models import User
 from django.db import models
 
-from surveyor.models import Group, Question
+from core.models import Group, Question
 
 """
 Django auto generates ID primary keys for each model, 
@@ -32,7 +32,7 @@ class Response(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
     question = models.ForeignKey(Question, on_delete=models.CASCADE)
     respondent = models.ForeignKey(Respondent, on_delete=models.SET_NULL, null=True)
-    value = models.SmallIntegerField(null=True)
+    value = models.FloatField(null=True)
     text = models.CharField(max_length=30, null=True)
     text_positive = models.BooleanField(null=True, default=None)
     date_time = models.DateTimeField()
